+++ examples/ssr-full/packages/pre-rendered-client/src/main.rs
@@ -1,4 +1,4 @@
-use ssr_example_app::hydrate_app;
+use ssr_full_app::hydrate_app;
 
 fn main() {
     hydrate_app()