+++ examples/ssr-full/packages/xtask/src/main.rs
@@ -2,7 +2,7 @@ use std::path::Path;
 
 use log::LevelFilter;
 use silkenweb::{dom::Dry, router, task};
-use ssr_example_app::app;
+use ssr_full_app::app;
 use xshell::Shell;
 use xtask_wasm::{
     anyhow::Result,