+++ packages/silkenweb/src/lib.rs
@@ -64,7 +64,7 @@
 //! [counter]: https://github.com/silkenweb/silkenweb/tree/main/examples/counter
 //! [routing]: https://github.com/silkenweb/silkenweb/tree/main/examples/router
 //! [tauri]: https://github.com/silkenweb/tauri-example
-//! [Server Side Rendering]: https://github.com/silkenweb/ssr-example
+//! [Server Side Rendering]: https://github.com/silkenweb/silkenweb/tree/main/examples/ssr-full
 //! [examples]: https://github.com/silkenweb/silkenweb/tree/main/examples
 //! [Declarative Shadow DOM]: https://web.dev/declarative-shadow-dom/
 use std::{cell::RefCell, collections::HashMap};